Changed dicionary size

+++ DataCompression/DataReader.cpp
@@ -49,12 +49,12 @@ void DataReader::readTextFile(std::ifstream &file)
 {
 	_config.word_bit_count = 8;
 	_config.data_order_type = ReadTypeEnum::LinearAscii;
-	_buffer.push_back(vector<uint16_t>());
+	_buffer.push_back(vector<uint32_t>());
 	char c;
 	while (file.get(c))
 	{
 		uint8_t byte = (uint8_t)c;
-		_buffer[0].push_back((uint16_t)byte);
+		_buffer[0].push_back((uint32_t)byte);
 	}
 }
 
@@ -62,7 +62,7 @@ void DataReader::readPgmFile(std::ifstream &file)
 {
 	_config.word_bit_count = 8;
 	_config.data_order_type = ReadTypeEnum::Linear;
-	_buffer.push_back(vector<uint16_t>());
+	_buffer.push_back(vector<uint32_t>());
 	//read and copy header info of pgm file//
 	int PGM_HEADER_LINES = 3;
 	int pgm_maxval = 255;
@@ -99,13 +99,13 @@ void DataReader::readPgmFile(std::ifstream &file)
 		_config.data_order_type = ReadTypeEnum::Split3Color;
 	}
 	//Get values to the buffer and reassemble as correct bitness//
-	_buffer.push_back(vector<uint16_t>());
+	_buffer.push_back(vector<uint32_t>());
 	char c;
 	if (_config.data_order_type = ReadTypeEnum::Linear)
 		while (file.get(c))
 		{
 			uint8_t byte = (uint8_t)c;
-			_buffer[1].push_back((uint16_t)byte);
+			_buffer[1].push_back((uint32_t)byte);
 		}
 	else if (_config.data_order_type = ReadTypeEnum::Split3Color)
 	{
@@ -121,7 +121,7 @@ void DataReader::readPgmFile(std::ifstream &file)
 		for (int i = 0; i < 3; ++i)
 			for (int j = 0; j < colors[i].size(); ++j)
 			{
-				_buffer[1].push_back((uint16_t)colors[i][j]);
+				_buffer[1].push_back((uint32_t)colors[i][j]);
 			}
 	}
 }
@@ -147,13 +147,13 @@ void DataReader::readLzwFile(std::ifstream &file)
 		 charHeader[i] = byte;
 	}
 	_config = compHeader;
-	_buffer.push_back(vector<uint16_t>());
+	_buffer.push_back(vector<uint32_t>());
 	//Decoding stage//
 	uint32_t lowerMask, higherMask;
 	uint32_t remainder = 0, shift = 0;
 	const uint8_t byteSize = 8;
-	uint16_t indexMask = 0;
-	uint16_t outputIndex = 0;
+	uint32_t indexMask = 0;
+	uint32_t outputIndex = 0;
 	int bitsToProcess;
 	//Reassemble 8 bit to output bit size//
 	for (int i = 0; i < _config.indx_bit_count; ++i)
@@ -174,7 +174,7 @@ void DataReader::readLzwFile(std::ifstream &file)
 			{
 				outputIndex = lowerMask & remainder;
 				outputIndex |= ((higherMask >> shift) & byte) << shift;
-				_buffer[0].push_back((uint16_t)outputIndex);
+				_buffer[0].push_back((uint32_t)outputIndex);
 				remainder = byte >> newDataBits;
 				shift = byteSize - newDataBits;
 				bitsToProcess -= _config.indx_bit_count;
@@ -192,7 +192,7 @@ LZWCompressHeader DataReader::getConfig()
 {
 	return _config;
 }
-vector<std::vector<uint16_t>>& DataReader::getBuffer()
+vector<std::vector<uint32_t>>& DataReader::getBuffer()
 {
 	return _buffer;
 }
